# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "neetpay API",
    "description": "Self-hosted crypto payment gateway. Pay without permission.",
    "version": "1.0.0",
    "contact": {
      "name": "neetpay",
      "url": "https://neetpay.io"
    }
  },
  "servers": [
    {
      "url": "https://neetpay.io/api/v1",
      "description": "Production"
    }
  ],
  "security": [
    { "apiKey": [] },
    { "bearerAuth": [] }
  ],
  "paths": {
    "/payment/create": {
      "post": {
        "operationId": "createPayment",
        "summary": "Create a payment",
        "description": "Generate a unique deposit address and payment link. Supports idempotency via the Idempotency-Key header. Accepts both sk_live_ (secret) and pk_live_ (publishable) keys.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": { "type": "string" },
            "description": "Unique key to prevent duplicate payments on retries"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePaymentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/CreatePaymentResponse" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key. Use sk_live_ (secret) for server-side calls, pk_live_ (publishable) for browser SDK. Publishable keys can only create payments."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token using your API key: Authorization: Bearer sk_live_..."
      }
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "description": "Standard response envelope for all v1 endpoints",
        "properties": {
          "data": { "description": "Response payload" },
          "message": { "type": "string" },
          "error": { "type": "null" },
          "status": { "type": "integer" },
          "version": { "type": "string", "example": "1.0" }
        }
      },
      "CreatePaymentRequest": {
        "type": "object",
        "required": ["amount", "payCurrencyKey"],
        "properties": {
          "amount": { "type": "number", "exclusiveMinimum": 0, "description": "Payment amount in fiat currency" },
          "currency": { "type": "string", "default": "USD", "description": "Fiat currency code (USD, EUR, GBP)" },
          "payCurrencyKey": { "type": "string", "description": "Crypto to accept: BTC, ETH, SOL, XMR, USDT-ERC20, etc." },
          "description": { "type": "string", "maxLength": 500 },
          "lifetimeMinutes": { "type": "integer", "minimum": 5, "maximum": 1440, "default": 30 },
          "callbackUrl": { "type": "string", "format": "uri", "description": "Per-payment webhook URL. Overrides account-level webhook URL for this payment." },
          "returnUrl": { "type": "string", "format": "uri", "description": "Redirect URL after successful payment" },
          "metadata": { "type": "object", "additionalProperties": true, "description": "Custom key-value data attached to the payment" }
        }
      },
      "CreatePaymentResponse": {
        "type": "object",
        "properties": {
          "trackId": { "type": "string", "example": "vp_m1abc_def456" },
          "payAddress": { "type": "string", "description": "Deposit address" },
          "chain": { "type": "string", "example": "bitcoin" },
          "payCurrency": { "type": "string", "example": "BTC" },
          "network": { "type": "string", "example": "mainnet" },
          "expectedPayAmount": { "type": "number", "nullable": true },
          "requiredConfirmations": { "type": "integer" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "explorerUrl": { "type": "string" }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "data": { "type": "object" },
          "message": { "type": "string" },
          "error": {
            "type": "object",
            "properties": {
              "type": { "type": "string" },
              "key": { "type": "string" },
              "message": { "type": "string" }
            }
          },
          "status": { "type": "integer" },
          "version": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      },
...
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": { "schema": { "type": "integer" }, "description": "Seconds until limit resets" }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
      }
    }
  },
...
}